Assignment9/requirements.txt
streamlit
torch
torchvision
Pillow
numpy==1.24.3 